symbiflow: fix check-env
