+++ fpgaperf.py
@@ -39,28 +39,6 @@ class NotAvailable:
     pass
 
 
-# https://stackoverflow.com/questions/377017/test-if-executable-exists-in-python
-def which(program):
-    def is_exe(fpath):
-        return os.path.isfile(fpath) and os.access(fpath, os.X_OK)
-
-    fpath, fname = os.path.split(program)
-    if fpath:
-        if is_exe(program):
-            return program
-    else:
-        for path in os.environ["PATH"].split(os.pathsep):
-            exe_file = os.path.join(path, program)
-            if is_exe(exe_file):
-                return exe_file
-
-    return None
-
-
-def have_exec(mybin):
-    return which(mybin) != None
-
-
 def print_stats(t):
     def print_section_header(title):
         print('')
